Add option for unsorted output

+++ src/main/java/datastructure/DupStats.java
@@ -1,10 +1,10 @@
 package main.java.datastructure;
 
 public class DupStats {
-	public int total = 0;
-	public int removed_reverse = 0;
-	public int removed_forward = 0;
-	public int removed_merged = 0;
+	public long total = 0;
+	public long removed_reverse = 0;
+	public long removed_forward = 0;
+	public long removed_merged = 0;
 
   public DupStats () {};
